removed different libraries and use of single binary
