+++ crates/router/src/core/payments.rs
@@ -19,7 +19,6 @@ use self::{
     flows::{ConstructFlowSpecificData, Feature},
     operations::{BoxedOperation, Operation},
 };
-use super::errors::StorageErrorExt;
 use crate::{
     core::errors::{self, RouterResponse, RouterResult},
     db::StorageInterface,
@@ -30,7 +29,6 @@ use crate::{
     types::{
         self, api,
         storage::{self, enums as storage_enums},
-        transformers::ForeignInto,
     },
     utils::OptionExt,
 };
@@ -541,6 +539,7 @@ pub fn should_call_connector<Op: Debug, F: Clone>(
     }
 }
 
+#[cfg(feature = "olap")]
 pub async fn list_payments(
     db: &dyn StorageInterface,
     merchant: storage::MerchantAccount,
@@ -551,11 +550,16 @@ pub async fn list_payments(
     let payment_intent =
         helpers::filter_by_constraints(db, &constraints, merchant_id, merchant.storage_scheme)
             .await
-            .map_err(|err| err.to_not_found_response(errors::ApiErrorResponse::PaymentNotFound))?;
+            .map_err(|err| {
+                errors::StorageErrorExt::to_not_found_response(
+                    err,
+                    errors::ApiErrorResponse::PaymentNotFound,
+                )
+            })?;
 
     let data: Vec<api::PaymentsResponse> = payment_intent
         .into_iter()
-        .map(ForeignInto::foreign_into)
+        .map(types::transformers::ForeignInto::foreign_into)
         .collect();
     Ok(services::ApplicationResponse::Json(
         api::PaymentListResponse {
