+++ crates/router/src/core/refunds.rs
@@ -476,6 +476,7 @@ pub async fn validate_and_create_refund(
 ///   If payment-id is not provided, lists the refunds associated with that particular merchant - to the limit specified,if no limits given, it is 10 by default
 
 #[instrument(skip_all)]
+#[cfg(feature = "olap")]
 pub async fn refund_list(
     db: &dyn db::StorageInterface,
     merchant_account: storage::merchant_account::MerchantAccount,
