+++ crates/router/src/db/refund.rs
@@ -56,6 +56,7 @@ pub trait RefundInterface {
         storage_scheme: enums::MerchantStorageScheme,
     ) -> CustomResult<storage_types::Refund, errors::StorageError>;
 
+    #[cfg(feature = "olap")]
     async fn filter_refund_by_constraints(
         &self,
         merchant_id: &str,
@@ -172,6 +173,7 @@ mod storage {
                 .into_report()
         }
 
+        #[cfg(feature = "olap")]
         async fn filter_refund_by_constraints(
             &self,
             merchant_id: &str,
@@ -179,6 +181,7 @@ mod storage {
             _storage_scheme: enums::MerchantStorageScheme,
             limit: i64,
         ) -> CustomResult<Vec<storage_models::refund::Refund>, errors::StorageError> {
+            #[cfg(feature = "olap")]
             let conn = pg_connection(&self.replica_pool).await;
             <storage_models::refund::Refund as storage_types::RefundDbExt>::filter_by_constraints(
                 &conn,
@@ -575,6 +578,7 @@ mod storage {
             }
         }
 
+        #[cfg(feature = "olap")]
         async fn filter_refund_by_constraints(
             &self,
             merchant_id: &str,
@@ -701,6 +705,7 @@ impl RefundInterface for MockDb {
         Err(errors::StorageError::MockDbError)?
     }
 
+    #[cfg(feature = "olap")]
     async fn filter_refund_by_constraints(
         &self,
         _merchant_id: &str,
