+++ crates/router/src/routes/payments.rs
@@ -5,7 +5,6 @@ use router_env::{instrument, tracing, Flow};
 use crate::{
     self as app,
     core::{errors::http_not_implemented, payments},
-    routes,
     services::{api, authentication as auth},
     types::api::{self as api_types, enums as api_enums, payments as payment_types},
 };
@@ -13,7 +12,7 @@ use crate::{
 #[instrument(skip_all, fields(flow = ?Flow::PaymentsCreate))]
 // #[post("")]
 pub async fn payments_create(
-    state: web::Data<routes::AppState>,
+    state: web::Data<app::AppState>,
     req: actix_web::HttpRequest,
     json_payload: web::Json<payment_types::PaymentsRequest>,
 ) -> impl Responder {
@@ -43,7 +42,7 @@ pub async fn payments_create(
 
 #[instrument(skip(state), fields(flow = ?Flow::PaymentsStart))]
 pub async fn payments_start(
-    state: web::Data<routes::AppState>,
+    state: web::Data<app::AppState>,
     req: actix_web::HttpRequest,
     path: web::Path<(String, String, String)>,
 ) -> impl Responder {
@@ -75,7 +74,7 @@ pub async fn payments_start(
 #[instrument(skip(state), fields(flow = ?Flow::PaymentsRetrieve))]
 // #[get("/{payment_id}")]
 pub async fn payments_retrieve(
-    state: web::Data<routes::AppState>,
+    state: web::Data<app::AppState>,
     req: actix_web::HttpRequest,
     path: web::Path<String>,
     json_payload: web::Query<payment_types::PaymentRetrieveBody>,
@@ -114,7 +113,7 @@ pub async fn payments_retrieve(
 #[instrument(skip_all, fields(flow = ?Flow::PaymentsUpdate))]
 // #[post("/{payment_id}")]
 pub async fn payments_update(
-    state: web::Data<routes::AppState>,
+    state: web::Data<app::AppState>,
     req: actix_web::HttpRequest,
     json_payload: web::Json<payment_types::PaymentsRequest>,
     path: web::Path<String>,
@@ -155,7 +154,7 @@ pub async fn payments_update(
 #[instrument(skip_all, fields(flow = ?Flow::PaymentsConfirm))]
 // #[post("/{payment_id}/confirm")]
 pub async fn payments_confirm(
-    state: web::Data<routes::AppState>,
+    state: web::Data<app::AppState>,
     req: actix_web::HttpRequest,
     json_payload: web::Json<payment_types::PaymentsRequest>,
     path: web::Path<String>,
@@ -197,7 +196,7 @@ pub async fn payments_confirm(
 #[instrument(skip_all, fields(flow = ?Flow::PaymentsCapture))]
 // #[post("/{payment_id}/capture")]
 pub async fn payments_capture(
-    state: web::Data<routes::AppState>,
+    state: web::Data<app::AppState>,
     req: actix_web::HttpRequest,
     json_payload: web::Json<payment_types::PaymentsCaptureRequest>,
     path: web::Path<String>,
@@ -228,7 +227,7 @@ pub async fn payments_capture(
 
 #[instrument(skip_all, fields(flow = ?Flow::PaymentsSessionToken))]
 pub async fn payments_connector_session(
-    state: web::Data<routes::AppState>,
+    state: web::Data<app::AppState>,
     req: actix_web::HttpRequest,
     json_payload: web::Json<payment_types::PaymentsSessionRequest>,
 ) -> impl Responder {
@@ -261,7 +260,7 @@ pub async fn payments_connector_session(
 
 #[instrument(skip_all)]
 pub async fn payments_response(
-    state: web::Data<routes::AppState>,
+    state: web::Data<app::AppState>,
     req: actix_web::HttpRequest,
     path: web::Path<(String, String, String)>,
 ) -> impl Responder {
@@ -294,7 +293,7 @@ pub async fn payments_response(
 #[instrument(skip_all, fields(flow = ?Flow::PaymentsCancel))]
 // #[post("/{payment_id}/cancel")]
 pub async fn payments_cancel(
-    state: web::Data<routes::AppState>,
+    state: web::Data<app::AppState>,
     req: actix_web::HttpRequest,
     json_payload: web::Json<payment_types::PaymentsCancelRequest>,
     path: web::Path<String>,
@@ -323,9 +322,10 @@ pub async fn payments_cancel(
 }
 
 #[instrument(skip_all, fields(flow = ?Flow::PaymentsList))]
+#[cfg(feature = "olap")]
 // #[get("/list")]
 pub async fn payments_list(
-    state: web::Data<routes::AppState>,
+    state: web::Data<app::AppState>,
     req: actix_web::HttpRequest,
     payload: web::Query<payment_types::PaymentListConstraints>,
 ) -> impl Responder {
@@ -344,7 +344,7 @@ pub async fn payments_list(
 
 async fn authorize_verify_select<Op>(
     operation: Op,
-    state: &routes::AppState,
+    state: &app::AppState,
     merchant_account: storage_models::merchant_account::MerchantAccount,
     req: api_models::payments::PaymentsRequest,
     auth_flow: api::AuthFlow,
