+++ crates/router/src/scheduler/workflows/payment_sync.rs
@@ -2,11 +2,9 @@ use router_env::logger;
 
 use super::{PaymentsSyncWorkflow, ProcessTrackerWorkflow};
 use crate::{
-    core::{
-        errors,
-        payments::{self as payment_flows, operations},
-    },
+    core::payments::{self as payment_flows, operations},
     db::{get_and_deserialize_key, StorageInterface},
+    errors,
     routes::AppState,
     scheduler::{consumer, process_data, utils},
     types::{
